리팩토링: AuditingFields 클래스를 추상 클래스로 변경
엔티티에서 상속을 해서 사용해야 하는 목적에 더 잘 맞게
`abstract` 키워드 추가

패키지 수정
