+++ src/main/java/com/fastcampus/projectboard/FastCampusProjectBoardApplication.java
@@ -1,4 +1,4 @@
-package fastcampus.projectboard;
+package com.fastcampus.projectboard;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
