+++ src/main/java/com/fastcampus/projectboard/domain/AuditingFields.java
@@ -1,4 +1,4 @@
-package fastcampus.projectboard.domain;
+package com.fastcampus.projectboard.domain;
 
 import lombok.Getter;
 import lombok.ToString;
@@ -18,7 +18,7 @@ import java.time.LocalDateTime;
 @ToString
 @EntityListeners(AuditingEntityListener.class)
 @MappedSuperclass
-public class AuditingFields {
+public abstract class AuditingFields {
     @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
     @CreatedDate
     @Column(nullable = false, updatable = false) private LocalDateTime createdAt; // 생성일시
