+++ src/test/java/com/fastcampus/projectboard/FastcampusProjectBoardApplicationTests.java
@@ -1,4 +1,4 @@
-package fastcampus.projectboard;
+package com.fastcampus.projectboard;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
